Function to update "Show more" button text and state
Added a function to update the "Show more" button text and state based on the number of remaining books to be displayed. The function sets the button text to "Show more (n)" where n is the number of remaining books, and disables the button if there are no more books to show.